adds selected on navigation bar buttons
