+++ FrontEnd/src/components/DefaultChatRoomWindow.tsx
@@ -1,4 +1,6 @@
 import "../styles/DefaultChatRoomLayout.css";
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faBolt } from '@fortawesome/free-solid-svg-icons';
 
 interface DefaultChatRoomProps {
     displayTab: string;
@@ -7,12 +9,12 @@ interface DefaultChatRoomProps {
 
 function DefaultChatRoomWindow(props: DefaultChatRoomProps) {
   return (
-    <div >
-        <img className="brainstormIcon" src="/src/assets/icon.svg"/>
-        <h3>Welcome to Brainstorm</h3>
+    <div className="DefaultChatRoomLayout">
+        <FontAwesomeIcon icon={faBolt} title="brainstorm icon" className="brainstormIcon"/>
+        <h2>Welcome to Brainstorm</h2>
         <p>Discuss and Vote ideas with your team</p>
         <div className="CreateChatRoomButton" style={props.displayTab === "Direct Message List" ? { display: "none" } : { display: "flex" }}>
-            <button onClick={props.handleFunction}>Create Chat Room</button>
+          <button onClick={props.handleFunction}>Create Chat Room</button>
         </div>
     </div>
   );
