+++ FrontEnd/src/styles/ChatList.css
@@ -31,7 +31,6 @@
   align-items: center;
   padding: 5px;
   font-weight: bold;
-  font-family: Arial, Helvetica, sans-serif;
   font-size: large;
   border-bottom: 1px solid white;
 }
@@ -91,6 +90,12 @@
   border-radius: 5%;
 }
 
+.ChatListContainer .chat-list .chats .chat-item.active {
+  background-color: rgb(29, 25, 75);
+  border-left: 1px solid #FFC436;
+  border-radius: 5px;
+}
+
 .ChatListContainer .chat-list .chats .chat-item .chat-details {
   width: 100%;
   margin-left: 10px;
@@ -122,18 +127,24 @@
 .ChatRoomHeader {
   display: flex;
   flex-direction: row;
-  column-gap: 70px;
-  font-weight: bold;
   font-family: Arial, Helvetica, sans-serif;
   font-size: large;
+  justify-content: space-between;
+  width: 100%;
 }
 
 .CreateChatRoomIcon {
-  width: fit-content;
   cursor: pointer;
-  color: black;
+  color: white;
   border: none;
-  background-color: #ffc436;
-  border-radius: 5px;
+  background-color: transparent;
   padding: 2px;
+  font-size: 18px;
+}
+
+.DefaultLayout {
+  justify-items: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
 }
