+++ FrontEnd/src/styles/ChatRoomWindow.css
@@ -69,7 +69,7 @@
 .WindowContainer .MemberListContainer {
     display: flex;
     flex-direction: column;
-    width: 15%;
+    width: 20%;
     min-width: 200px;
-    background-color: #721f1f;
+    border-left: 1px solid white;
 }
