+++ FrontEnd/src/styles/DefaultChatRoomLayout.css
@@ -1,35 +1,30 @@
 .DefaultChatRoomLayout {
+    position: relative;
     display: flex;
     flex-direction: column;
-    /* align-items: center; */
+    align-items: center;
     justify-content: center;
-    margin: 0 40%;
+    font-family: Arial, Helvetica, sans-serif;
+    width: 100%;
+    height: 100%;
 }
 
 .DefaultChatRoomLayout .brainstormIcon {
-    color: rgb(255,172,51, 0.1);
-    size: 10px;
+    font-size: 250px;
+    color: rgba(232, 232, 232, 0.6);
 }
 
-.DefaultChatRoomLayout h3 {
-    font-size: 30px;
-    font-family: Arial, Helvetica, sans-serif;
-    width: 100%;
+.DefaultChatRoomLayout h2 {    
     margin-top: 50px;
     color: white;
-    align-items: center;
-    justify-content: center;
 }
 
 .DefaultChatRoomLayout p {
     font-size: large;
-    font-family: Arial, Helvetica, sans-serif;
-    width: 100%;
     color: rgb(183, 181, 181);
 }
 
 .CreateChatRoomButton {
-    margin: 0 20%;
     margin-top: 30px;
 }
 
@@ -49,5 +44,4 @@
 
 .CreateChatRoomButton button:hover {
     background-color: #e6ab22;
-    color: black;
 }
