+++ FrontEnd/src/styles/MemberList.css
@@ -4,7 +4,7 @@
   position: relative;
   width: 100%;
   height: 100%;
-  background-color: #0174BE;
+  background-color: #0c356a;
   color: white;
   display: flex;
   flex-direction: column;
@@ -16,8 +16,11 @@
 .member-list .offline-members h2 {
   position: relative;
   font-size: 22px;
+  font-family: Arial, Helvetica, sans-serif;
   display: flex;
   justify-content: start;
+  border-bottom: 1px solid white;
+  padding: 5px;
 }
 
 .member-list ul {
@@ -36,4 +39,9 @@
 
 .member-list .online-members li:hover {
   background-color: transparent;
+}
+
+.members {
+  position: relative;
+  margin: 5px;
 }
